Ajuste de estrutura da model Assistido (Não testado)
