+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/controllers/AssistidoController.java
@@ -1,7 +1,7 @@
 package com.uniprojecao.fabrica.gprojuridico.controllers;
 
 import com.uniprojecao.fabrica.gprojuridico.domains.Autocomplete.AssistidoAutocomplete;
-import com.uniprojecao.fabrica.gprojuridico.dto.assistido.AssistidoDTO;
+import com.uniprojecao.fabrica.gprojuridico.domains.assistido.Assistido;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.AssistidoMinDTO;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.AtendimentoVinculadoAssistidoDTO;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.ProcessoVinculado;
@@ -24,7 +24,7 @@ public class AssistidoController {
     private AssistidoService service;
 
     @PostMapping
-    public ResponseEntity<AssistidoDTO> insert(@RequestBody @Valid AssistidoDTO data) {
+    public ResponseEntity<Assistido> insert(@RequestBody @Valid Assistido data) {
         var result = service.insert(data);
         var id = result.getCpf();
         return ResponseEntity.created(createUri(id)).body(result);
@@ -74,8 +74,8 @@ public class AssistidoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<AssistidoDTO> findById(@PathVariable String id) {
-        AssistidoDTO result = service.findById(id);
+    public ResponseEntity<Assistido> findById(@PathVariable String id) {
+        Assistido result = service.findById(id);
         return ResponseEntity.ok(result);
     }
