+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/domains/Filiacao.java
@@ -1,0 +1,20 @@
+package com.uniprojecao.fabrica.gprojuridico.domains;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Filiacao {
+    @NotBlank
+    @Size(min = 3, max = 60)
+    private String mae;
+
+    @NotBlank
+    @Size(min = 3, max = 60)
+    private String pai;
+}
