+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/domains/assistido/AssistidoTrabalhista.java
@@ -1,6 +1,9 @@
 package com.uniprojecao.fabrica.gprojuridico.domains.assistido;
 
 import com.uniprojecao.fabrica.gprojuridico.domains.Endereco;
+import com.uniprojecao.fabrica.gprojuridico.domains.Filiacao;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,8 +13,13 @@ import java.util.Map;
 @NoArgsConstructor
 @Data
 public class AssistidoTrabalhista extends Assistido {
+    @NotNull
     private Ctps ctps;
+
+    @NotBlank
     private String pis;
+
+    @NotNull
     private Boolean empregadoAtualmente;
 
     public AssistidoTrabalhista(String nome, String rg, String cpf, String nacionalidade, String escolaridade, String estadoCivil, String profissao, String telefone, String email, Filiacao filiacao, String remuneracao, Map<String, Endereco> endereco, Ctps ctps, String pis, Boolean empregadoAtualmente) {
