+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/services/AssistidoService.java
@@ -1,9 +1,9 @@
 package com.uniprojecao.fabrica.gprojuridico.services;
 
 import com.uniprojecao.fabrica.gprojuridico.domains.Autocomplete.AssistidoAutocomplete;
+import com.uniprojecao.fabrica.gprojuridico.domains.assistido.Assistido;
 import com.uniprojecao.fabrica.gprojuridico.domains.assistido.AssistidoCivil;
 import com.uniprojecao.fabrica.gprojuridico.domains.assistido.AssistidoTrabalhista;
-import com.uniprojecao.fabrica.gprojuridico.dto.assistido.AssistidoDTO;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.AssistidoMinDTO;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.AtendimentoVinculadoAssistidoDTO;
 import com.uniprojecao.fabrica.gprojuridico.dto.min.ProcessoVinculado;
@@ -17,8 +17,6 @@ import java.util.List;
 import java.util.Map;
 
 import static com.uniprojecao.fabrica.gprojuridico.services.utils.Constants.ASSISTIDOS_COLLECTION;
-import static com.uniprojecao.fabrica.gprojuridico.services.utils.Utils.ModelMapper.toDto;
-import static com.uniprojecao.fabrica.gprojuridico.services.utils.Utils.ModelMapper.toEntity;
 import static com.uniprojecao.fabrica.gprojuridico.services.utils.Utils.filterValidKeys;
 import static com.uniprojecao.fabrica.gprojuridico.services.utils.Utils.initFilter;
 import static java.lang.Integer.parseInt;
@@ -33,9 +31,9 @@ public class AssistidoService extends BaseService {
         super(collectionName);
     }
 
-    public AssistidoDTO insert(AssistidoDTO dto) {
-        BaseRepository.save(collectionName, dto.getCpf(), toEntity(dto));
-        return dto;
+    public Assistido insert(Assistido data) {
+        BaseRepository.save(collectionName, data.getCpf(), data);
+        return data;
     }
 
     public List<AssistidoMinDTO> findAll(String limit, String field, String filter, String value) {
@@ -58,9 +56,8 @@ public class AssistidoService extends BaseService {
                 initFilter("assistidoId", "EQUAL", id));
     }
 
-    public AssistidoDTO findById(String id) {
-        var assistido = repository.findById(id);
-        return (assistido != null) ? toDto(assistido) : null;
+    public Assistido findById(String id) {
+        return repository.findById(id);
     }
 
     public void update(String id, Map<String, Object> data, String clazz) {
